Some Changes
add exception class, modifying delete,create,get services and employee entity

+++ src/main/java/com/example/humanResourcesProject/service/EmployeeDeleteService.java
@@ -1,6 +1,7 @@
 package com.example.humanResourcesProject.service;
 
 import com.example.humanResourcesProject.entity.Employee;
+import com.example.humanResourcesProject.exception.EmployeeNotFoundException;
 import com.example.humanResourcesProject.repository.EmployeeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,6 +14,13 @@ public class EmployeeDeleteService {
     EmployeeRepository employeeRepository;
 
     public void deleteEmployee(int id){
-        employeeRepository.deleteById(id);
+        Optional<Employee> employee = employeeRepository.findById(id);
+
+        if (!employee.isPresent()) {
+            throw new EmployeeNotFoundException(id);
+        } else {
+            employeeRepository.deleteById(id);
+        }
+
     }
 }
